refactor(teams): split runner execution.rs into dispatch and lifecycle modules [OPE-433]
Decompose the 501-line execution.rs into focused submodules:
- dispatch.rs: provider-chain fallback logic and all run variants (run, run_structured, run_streaming, run_with_events)
- lifecycle.rs: session accessors, state management, history seeding, and schema configuration
