+++ crates/opengoose-teams/src/runner/mod.rs
@@ -1,5 +1,6 @@
 mod construction;
-mod execution;
+mod dispatch;
+mod lifecycle;
 pub(crate) mod output;
 pub(crate) mod types;
